requirements-dev.txt
paddlepaddle>=2.4.2
torch>=1.12.0
pre-commit
pytest
parameterized
pytest-cov
regex
pytest-xdist
torchvision
